Enhance static site generation and configuration updates
- Added `generateStaticParams` function to multiple pages for improved static generation with support for 'ua' and 'en' locales.
- Updated `next.config.ts` to enable image optimization and set output to 'export'.
- Introduced new export script in `package.json` for static site export.
- Removed unused CSS file related to blog articles.
- Refactored blog page to utilize a new component for better structure and maintainability.
