+++ next.config.ts
@@ -1,8 +1,9 @@
-import type { NextConfig } from 'next';
+import type { NextConfig } from 'next'; // eslint-disable-line @typescript-eslint/no-unused-vars
 
 
 const nextConfig = {
   images: {
+    unoptimized: true,
     remotePatterns: [
       {
         protocol: 'https',
@@ -11,6 +12,7 @@ const nextConfig = {
       },
     ],
   },
+  output: 'export',
 };
 
 export default nextConfig;
