+++ src/app/[locale]/education/page.tsx
@@ -13,6 +13,13 @@ export const metadata = {
   description: "Ми підбираємо надійних нянь, яким можна довірити вашу дитину.",
 };
 
+export async function generateStaticParams() {
+  return [
+    { locale: 'ua' },
+    { locale: 'en' },
+  ];
+}
+
 export default async function Education({ params }: { params: Promise<{ locale: string }> }) {
   const { locale } = await params;
   const translationPath = path.resolve(process.cwd(), `public/locales/${locale}/common.json`);
