+++ src/app/[locale]/vacation/page.tsx
@@ -2,6 +2,13 @@ import fs from 'fs/promises';
 import path from 'path';
 import { VacationPage } from '@/components/VacationPage/VacationPage';
 
+export async function generateStaticParams() {
+  return [
+    { locale: 'ua' },
+    { locale: 'en' },
+  ];
+}
+
 export default async function Vacations({ params }: { params: Promise<{ locale: string }> }) {
   const { locale } = await params;
   const translationPath = path.resolve(process.cwd(), `public/locales/${locale}/common.json`);
